reverse dependency
https://github.com/CraneStation/wasmtime/pull/186#discussion_r302780035
